+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` w
 
 - Breaking: Replace Kraken Game/Title setting API calls with Helix ones. (#1001)
 - Minor: Bot now provides more helpful error message when `!clip` is used in a channel with clips disabled. (#1091)
+- Minor: Added custom message options to the LastFM module. (#1090)
+- Minor: Added customizable cooldowns to the LastFM module. (#1090)
+- Minor: Added an online only option to the LastFM module. (#1090)
 - Minor: Added option to select your referred gambling command. (#1067)
 - Minor: The `sub_only` command option will now show in the `!debug command` command. (#1027)
 - Minor: The permaban module will now ban immediately on command use (#1014)